+++ lib/gitolite_recycle.rb
@@ -17,11 +17,6 @@ module GitHosting
 			return GitHosting.logger
 		end
 
-                # Repository base path (relative to git user home directory)
-                def self.repository_base
-                	Setting.plugin_redmine_git_hosting['gitRepositoryBasePath']
-                end
-
                 # Recycle bin base path (relative to git user home directory)
                 def self.recycle_bin
                 	Setting.plugin_redmine_git_hosting['gitRecycleBasePath'] || RECYCLE_BIN_IF_UNDEF
@@ -33,9 +28,10 @@ module GitHosting
                 end
                   
                 # Scan through the recyclebin and delete files older than 'preserve_time' minutes
-                # This can fail silently if, for instance, the recycle_bin doesn't exist.  That is the intended behavior.
           	def self.delete_expired_files
-                	result = %x[#{GitHosting.git_user_runner} 'find #{recycle_bin} -type d -regex '.*\.git' -cmin +#{preserve_time} -prune -print'].chomp.split("\n")
+                	return unless GitHosting.file_exists?(recycle_bin)
+
+                	result = %x[#{GitHosting.git_user_runner} find '#{recycle_bin}' -type d -regex '.*\.git' -cmin +#{preserve_time} -prune -print].chomp.split("\n")
                 	if result.length > 0
                         	logger.warn "Garbage-collecting expired file#{(result.length != 1) ? "s" : ""} from recycle bin:"
                         	result.each do |filename| 
@@ -53,21 +49,27 @@ module GitHosting
                 end
 
                 def self.move_repository_to_recycle repo_name
-                	repo_path = File.join(repository_base, repo_name)
-                  	new_path = File.join(recycle_bin,"#{Time.now.to_i.to_s}#{TRASH_DIR_SEP}#{name_to_recycle_name(repo_name)}")                	
+                	# Only bother if actually exists!
+                	return unless GitHosting.git_repository_exists?(repo_name)
+                	
+                	repo_path = GitHosting.repository_path(repo_name)
+                  	new_path = File.join(recycle_bin,"#{Time.now.to_i.to_s}#{TRASH_DIR_SEP}#{name_to_recycle_name(repo_name)}.git")                	
                 	begin
                         	GitHosting.shell %[#{GitHosting.git_user_runner} mkdir -p '#{recycle_bin}']
-                        	GitHosting.shell %[#{GitHosting.git_user_runner} mv '#{repo_path}.git' '#{new_path}.git']
-                        	logger.warn "  Moving '#{repo_name}.git' from gitolite repository => '#{new_path}.git'.  Will remain for at least #{preserve_time/60.0} hours"
+                        	GitHosting.shell %[#{GitHosting.git_user_runner} chmod 770 '#{recycle_bin}']
+                        	GitHosting.shell %[#{GitHosting.git_user_runner} mv '#{repo_path}' '#{new_path}']
+                        	logger.warn "  Moving '#{repo_name}' from gitolite repository => '#{new_path}'.  Will remain for at least #{preserve_time/60.0} hours"
                           	# If any empty directories left behind, try to delete them.  Ignore failure.
                           	old_prefix = repo_name[/.*?(?=\/)/] # Top-level old directory without trailing '/'
                           	if old_prefix
-                                  	repo_subpath = File.join(repository_base, old_prefix)
+                                  	repo_subpath = File.join(GitHosting.repository_base, old_prefix)
                                 	result = %x[#{GitHosting.git_user_runner} find '#{repo_subpath}' -type d ! -regex '.*\.git/.*' -empty -depth -delete -print].chomp.split("\n")
                                 	result.each { |dir| logger.warn "  Removing empty repository subdirectory: #{dir}"}
                                 end
+                        	return true
                         rescue
                         	logger.error "Attempt to move repository '#{repo_name}.git' to recycle bin failed"
+                        	return false
                         end
                 end
 
@@ -77,16 +79,16 @@ module GitHosting
                   	files = %x[#{GitHosting.git_user_runner} find '#{recycle_bin}' -type d -regex '#{myregex}' -prune].chomp.split("\n").sort {|x,y| y <=> x }
                 	if files.length > 0
                         	# Found something!
-                          	logger.warn "Restoring '#{repo_name}.git' to gitolite repository from recycle bin (#{files.first})"
+                          	logger.warn "  Restoring '#{repo_name}.git' to gitolite repository from recycle bin (#{files.first})"
                         	begin
                                 	prefix = repo_name[/.*(?=\/)/] # Complete directory path (if exists) without trailing '/'
                         		if prefix
-                                        	repo_prefix = File.join(repository_base, prefix)
+                                        	repo_prefix = File.join(GitHosting.repository_base, prefix)
                                         	# Has subdirectory.  Must reconstruct directory
                                         	GitHosting.shell %[#{GitHosting.git_user_runner} mkdir -p '#{repo_prefix}']
                                         end
-                                	repo_path = File.join(repository_base, repo_name)
-                                	GitHosting.shell %[#{GitHosting.git_user_runner} mv '#{files.first}' '#{repo_path}.git']
+                                	repo_path = GitHosting.repository_path(repo_name)
+                                	GitHosting.shell %[#{GitHosting.git_user_runner} mv '#{files.first}' '#{repo_path}']
 
                                 	# Optionally remove recycle_bin (but only if empty).  Ignore error if non-empty
                         		%x[#{GitHosting.git_user_runner} rmdir #{recycle_bin}]
